change totalSize type from size_t to ull
totalSize is compared against a value of 50 billion.  On 32 bits cpu
architectures, a size_t is an unsigned int effectively 32 bits wide,
per definition of size_t.  However the integer would have overflown
already at 5 billion.  On 64 bits cpu architecture, the size_t
matches effectively the type of an unsigned long long, so aligning
to that type avoids the overflow.

The issue has been caught initially by Adrian Bunk, from the Debian QA
group, on Debian bug #989855[1].

[1]: https://bugs.debian.org/989855

+++ ntcard.cpp
@@ -424,7 +424,7 @@ main(int argc, char** argv)
 			inFiles.push_back(file);
 	}
 
-	size_t totalSize = 0;
+	unsigned long long int totalSize = 0;
 	for (unsigned file_i = 0; file_i < inFiles.size(); ++file_i)
 		totalSize += getInf(inFiles[file_i].c_str());
 	if (totalSize < 50000000000)
